covars working for both kinds of logregression. and both give identical results now
